@@ -23,8 +23,6 @@ classifiers = [
     "Environment :: Plugins",
     "Intended Audience :: Science/Research",
     "License :: OSI Approved :: MIT License",
-    "Programming Language :: Python :: 3.7",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
@@ -41,10 +39,10 @@ keywords = [
     "cache",
 ]
 urls = {Homepage = "https://aiida-testing.readthedocs.io/"}
-requires-python = ">=3.7"
+requires-python = ">=3.9"
 # Note the dependency on setuptools due to pkg_resources
 dependencies = [
-    "aiida-core>=1.0.0,<2.6",
+    "aiida-core>=2.0,<2.6",
     "pytest>=7.0",
     "voluptuous~=0.12",
     "setuptools",
@@ -54,7 +52,7 @@ dependencies = [
 docs = ["sphinx", "sphinx-rtd-theme"]
 tests = [
     "pgtest~=1.3.1",
-    "aiida-diff",
+    "aiida-diff~=2.0.0",
     "pytest-datadir",
     "pytest-mock",
     "pytest-cov>=4.0",
@@ -136,16 +134,16 @@ ignore_missing_imports = true
 [tool.tox]
 legacy_tox_ini = """
 [tox]
-envlist = py38
+envlist = py39
 
 [testenv]
 usedevelop = true
 
-[testenv:py{37,38,39,310,311}]
+[testenv:py{39,310,311,312}]
 extras = tests
 commands = pytest {posargs}
 
-[testenv:pre-commit-py{38,39,310,311}]
+[testenv:pre-commit-py{39,310,311,312}]
 extras = 
     pre_commit
 commands = pre-commit {posargs:run --all}
