@@ -7,6 +7,6 @@ to enable AiiDA - level caching.
 from ._fixtures import *
 
 __all__ = (
-    "pytest_addoption", "absolute_archive_path", "enable_archive_cache", "liberal_hash",
-    "archive_cache_forbid_migration", "archive_cache_overwrite"
+    "absolute_archive_path", "archive_cache_forbid_migration", "archive_cache_overwrite",
+    "enable_archive_cache", "liberal_hash", "pytest_addoption"
 )
